refactor: move healthstatus into pkg
